tenantbackend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
celery==5.4.0
charset-normalizer==3.1.0
click==8.1.6
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.2
Django==5.0.6
django-cors-headers==4.7.0
django-tenant-schemas==1.12.0
django-tenants==3.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
psycopg2-binary==2.9.10

PyJWT==2.9.0
pymongo==4.11.1

pyotp==2.9.0
pyparsing==3.0.9

pytz==2023.3
PyYAML==6.0.2  # For YAML file processing
requests==2.30.0
sqlparse==0.5.0
urllib3==2.0.2
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.2